requirements.txt
paho-mqtt==1.5.1
ConfigArgParse==1.5.3
markovify==0.9.3
cachetools==4.2.4
